+++ Clove/.gitignore
@@ -1,18 +1,37 @@
-*.class
-*.war
-*.ear
+### Gradle ###
+.gradle
+build/
+!gradle/wrapper/gradle-wrapper.jar
+!**/src/main/**/build/
+!**/src/test/**/build/
 
-/bin/
-.metadata
-.settings
-.recommenders
+### STS ###
+.apt_generated
 .classpath
+.factorypath
 .project
+.settings
+.springBeans
+.sts4-cache
 
-**/src/parser/ast/*.java
-**/src/parser/ast/*.jj
+### IntelliJ IDEA ###
+.idea
+*.iws
+*.iml
+*.ipr
+out/
+!**/src/main/**/out/
+!**/src/test/**/out/
 
-/code-snippets/*.txt
-/example-programs/*.txt
-/example-programs/*.json
-/tests/*.txt
+### NetBeans ###
+/nbproject/private/
+/nbbuild/
+/dist/
+/nbdist/
+/.nb-gradle/
+
+### VS Code ###
+.vscode/
+
+### Clove-lang ###
+**/parser/ast/*
