+++ Clove/build.gradle
@@ -1,0 +1,24 @@
+plugins {
+    id 'application'
+}
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.7.2'
+}
+
+application {
+    mainClass = 'dev.amrw.clovelang.CloveMain'
+}
+
+run {
+    // Let `stdin` through.
+    standardInput = System.in
+}
+
+tasks.named('test') {
+    useJUnitPlatform()
+}
