--- Clove/src/Clove.java
@@ -1,5 +1,0 @@
-public class Clove {
-	public static void main(String[] args) {
-		interpreter.Interpreter.main(args);
-	}
-}
