+++ Clove/src/main/java/dev/amrw/clovelang/CloveMain.java
@@ -1,0 +1,7 @@
+package dev.amrw.clovelang;
+
+public class CloveMain {
+	public static void main(String[] args) {
+		dev.amrw.clovelang.interpreter.Interpreter.main(args);
+	}
+}
