+++ Clove/tests/.gitignore
@@ -1,0 +1,2 @@
+test33.txt
+fileFuncTest.txt
